requirements.txt
pytesseract
pillow
googletrans==4.0.0rc1
flask
transliterate
